New spell: Transdimensional Blitz (L5 Translocations)
Transdimensional Blitz launches melee attacks with against 2-4 random
enemies in LOS, depending on spellpower. It's themed as opening one
"in" portal and a bunch of tiny "out" portals, letting you hit many
enemies with a single attack.

This spell is intended to give melee-focused characters more reason to
consider investing in spellcasting, It's fun to try to skill toward a
'hybrid' build, and a spell that mostly scales off your melee attack,
rather than your spellpower, helps encourage that. This being a L5
spell helps avoid it being a truly free 'gimme', and allows it to hit
many enemies at once.

The lack of targeter should hopefully avoid it being annoying to use,
by making it just a single press to cast. It also gives the player
more reason to keep normally hitting things in melee sometimes,
when it's important to kill a specific target.

Cleaving doesn't work with this spell since it seems overly confusing,
and since reach and riposte don't work with it either. (Thus, it's
'fair'.)

Rename and re-level Projected Weapon

Give it Air school so that it's not a pure-tloc damage spell, and
call it Vortex of Steel instead, since that seems cooler. Not renaming
the enums until I finalize this.

Don't allow sleepstabbing with Proj Weapon

A bit too easy.